More test changes
